[vim] Use abort for all Vimscript functions
I encountered a bug in :Help because a Vimscript function was not marked
with 'abort'. I think I learned my lesson, and I now understand the
recommentation of some style guides [1]: all functions should be marked
with 'abort'.

Mark all Vimscript functions with 'abort' to avoid latent bugs which are
hard to discover in tests.

[1] https://google.github.io/styleguide/vimscriptguide.xml?showone=Functions#Functions