[vim] Fix test_messages since Vim v8.2.0636
Vim version 8.2.0636 introduced a patch which causes the 'Messages
maintainer: Bram Moolenaar <Bram@vim.org>' message to no longer appear
in some cases. Since that patch, Vim uses proper locale lookup instead
of blindingly trusting the LANG environment variable. If the LANG
variable is malformed or if the locale is not installed, Vim no longer
shows the 'Messages maintainer' message.

Try to detect whether or not 'Messages maintainer' will be printed using
'helplang, which is set in a way similar to the messages language.

For Vim versions older than 8.2.0636, this commit should not change
behavior.